Carga de inicialización del schema y data en H2

+++ src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     nombre VARCHAR(100) NOT NULL,
-    precio DOUBLE PRECISION NOT NULL,
+    precio DOUBLE NOT NULL,
     stock INT NOT NULL
 );
